refactor: extract AppState support-navigation extension (#568, #430)
First slice of #430 (AppState decomposition). Carves the 7 support-
navigation methods (openAbout, openChangelog, openGitHubRepository,
openDocumentation, openIssueReporter, openSupportDevelopment,
openSupportDonationPage) out of the 1,425-line AppState.swift and into
their own extension file.

Moves lines 894-920 (27 lines, 731 bytes, SHA-256 030f6573239f25b8b8e396b5612b1fcdc45707c48bfb6bc59f135e08d5ec43a7)
into a new Sources/BugNarrator/AppState+SupportNavigation.swift as
`extension AppState { ... }`. Byte-identical bodies. The extension
inherits @MainActor from AppState's class-level annotation (SE-0316,
verified by codex spec review) — do NOT add @MainActor to the
extension itself.

Endpoints verified before extraction (`sed -n '894p'` = `func openAbout()
{`; `'920p'` = `}` of openSupportDonationPage). Byte-diff proof against
origin/main lines 894-920 is silent (exit 0); SHA-256 of the extracted
block in the new file matches the origin SHA. 662 unit tests pass
unchanged; docs-sync validation passes.

The mixed delegation pattern is preserved: 3 methods call
appUtilityActions.<method>() directly (openAbout, openChangelog,
openSupportDevelopment); 4 go through
appUtilityActionPresenter.present(appUtilityActions.<method>())
(openGitHubRepository, openDocumentation, openIssueReporter,
openSupportDonationPage). Every SwiftUI call site (menu bar, About
window, support views) continues to resolve appState.openAbout() etc.
identically — the extension is same-module and inherits AppState's
isolation.

New file imports Foundation only (matches house style; codex confirmed
the method bodies name only same-module members — no Foundation types
are strictly required, but Foundation is the conventional floor).

Pre-build spec reviewed by codex in a fresh session (3 findings folded
in: SHA/byte-count checksum, Xcode target membership via xcodegen, tight
imports).
