+++ Sources/BugNarrator/AppState.swift
@@ -891,34 +891,6 @@ final class AppState: ObservableObject {
         applicationTerminationController.applicationShouldTerminate()
     }
 
-    func openAbout() {
-        appUtilityActions.openAbout()
-    }
-
-    func openChangelog() {
-        appUtilityActions.openChangelog()
-    }
-
-    func openGitHubRepository() {
-        appUtilityActionPresenter.present(appUtilityActions.openGitHubRepository())
-    }
-
-    func openDocumentation() {
-        appUtilityActionPresenter.present(appUtilityActions.openDocumentation())
-    }
-
-    func openIssueReporter() {
-        appUtilityActionPresenter.present(appUtilityActions.openIssueReporter())
-    }
-
-    func openSupportDevelopment() {
-        appUtilityActions.openSupportDevelopment()
-    }
-
-    func openSupportDonationPage() {
-        appUtilityActionPresenter.present(appUtilityActions.openSupportDonationPage())
-    }
-
     func openMicrophonePrivacySettings() {
         appUtilityActionPresenter.present(appUtilityActions.openMicrophonePrivacySettings())
     }
